Tyler's teaching notes, first commit
Project management and RMarkdown are taken and directly from Software Carpentry,
with some very light editing. I'll make a few more changes to 'localize' them,
and I'm not sure if I want to start with these, or do the plotting first, and
then come back to this.

The plotting lesson is very much a work in progress, mostly converting a
previous lesson plan into RMarkdown, and adding more hands-on exercises in place
of Tyler's brilliant 'look at me code!' demonstrations ;)